refactor(praxis-card): structural content slots + pickVariant dispatch
PraxisCard already dispatched per faction, but every archetype wrapped one
monolithic PraxisContent whose only per-faction variance was color overrides
(and a dead bodyStyle prop), so the cards read flat below the chrome (SPEC §1
#2). Breaks the content into composable structural slots (PraxisTitle,
PraxisTaskLink, PraxisByline) in praxisCard/shared.tsx, with PraxisContent now a
byte-identical default composition an archetype can rearrange. Extracts the
optimistic-moderation state into usePraxisCard, and replaces the slug switch
with a Record + pickVariant. DOM is unchanged for every faction; the alias rule
now routes albescent/aged_out task factions (which do not occur in practice) to
the ua card rather than the generic fallback.
